requirements.txt
Django==2.0.6
django-crispy-forms==1.7.2
django-filter==2.0.0
djangorestframework==3.8.2
Markdown==2.6.11
Pillow==5.2.0
pytz==2018.5
